chore: course type 정리
