+++ src/apis/course/types.ts
@@ -1,4 +1,4 @@
-import { TPlace } from "../../pages/Course/ShareCourse/Share/_components/ChatInterface";
+import { TPlace } from "../../pages/Course/ShareCourse/share.types";
 import { TShareCoursePlace } from "./index";
 import { VoteLimitTimeMap } from "../../pages/Course/ShareCourse/Share/constants";
 export type TVoteLimitTime = keyof typeof VoteLimitTimeMap;
