+++ src/pages/Course/ShareCourse/ShareCorsePage.tsx
@@ -4,12 +4,9 @@ import List from "./List";
 import Share from "./Share";
 import CommonContainer from "../../../components/Layout/CommonContainer";
 import Modal from "../../../components/Modal/core";
-import { ShareCourseListSchema } from "./schema";
-import * as z from "zod";
 import { DefaultPlace } from "../constants";
 import { ICourseInfo } from "../../../apis/course/types";
-
-export type ShareCourseData = z.infer<typeof ShareCourseListSchema>;
+import { ShareCourseData } from "./share.types";
 
 export type TTripInfo = ICourseInfo & {
   roomId: string;
